Button created for Researchers to view existing proposals
"My Proposals" button was created and it will only displays if the user logged in is a researcher, otherwise the home page will keep the current layout. Once the button is clicked, it will lead to a page displays all proposals submitted by current user. Currently the page is blank and it will be developed in next commit.